Refactorizacion de codigo
Problemas en DigitalOcean

+++ database/migrations/2022_12_29_042530_create_permission_tables.php
@@ -52,7 +52,7 @@ class CreatePermissionTables extends Migration
 
         Schema::create($tableNames['model_has_permissions'], function (Blueprint $table) use ($tableNames, $columnNames, $teams) {          
             
-             $table->unsignedBigInteger('id'); // agrego yo por problemas en digitalOcean
+            $table->increments('id'); // Agrego esta linea por error en laravel ocean.
             
             $table->unsignedBigInteger(PermissionRegistrar::$pivotPermission);
 
@@ -81,7 +81,7 @@ class CreatePermissionTables extends Migration
 
         Schema::create($tableNames['model_has_roles'], function (Blueprint $table) use ($tableNames, $columnNames, $teams) {
                   
-            $table->unsignedBigInteger('id'); // agrego yo por problemas en digitalOcean
+            $table->increments('id'); // Agrego esta linea por error en laravel ocean.
 
             $table->unsignedBigInteger(PermissionRegistrar::$pivotRole);
 
@@ -107,7 +107,7 @@ class CreatePermissionTables extends Migration
 
         Schema::create($tableNames['role_has_permissions'], function (Blueprint $table) use ($tableNames) {   
             
-            $table->unsignedBigInteger('id'); // agrego yo por problemas en digitalOcean
+            $table->increments('id'); // Agrego esta linea por error en laravel ocean.
 
             $table->unsignedBigInteger(PermissionRegistrar::$pivotPermission);
             $table->unsignedBigInteger(PermissionRegistrar::$pivotRole);
